+++ stories/hooks/useSpringBounceStop.story.lua
@@ -6,9 +6,9 @@ local GuiService = game:GetService("GuiService")
 local Roact = require(ReplicatedStorage.Packages.Roact)
 local Hooks = require(ReplicatedStorage.Packages.Hooks)
 local RoactSpring = require(ReplicatedStorage.Packages.RoactSpring)
+local CircleButton = require(script.Parent.Parent.components.CircleButton)
 
 local e = Roact.createElement
-
 local STOP_AFTER_SECONDS = 3
 
 local function Button(props, hooks)
@@ -18,14 +18,8 @@ local function Button(props, hooks)
         }
     end)
 
-	return e("TextButton", {
-        AnchorPoint = Vector2.new(0.5, 0.5),
+	return e(CircleButton, {
         Position = styles.position,
-		Size = UDim2.fromOffset(150, 150),
-        BackgroundColor3 = Color3.fromRGB(255, 255, 255),
-        AutoButtonColor = false,
-        Text = "",
-
         [Roact.Event.Activated] = function()
             api.start({
                 to = { position = UDim2.fromScale(0.5, 0.8) },
@@ -34,18 +28,7 @@ local function Button(props, hooks)
             task.wait(STOP_AFTER_SECONDS)
             api.stop()
         end
-	}, {
-        UICorner = e("UICorner", {
-            CornerRadius = UDim.new(1, 0),
-        }),
-        UIGradient = e("UIGradient", {
-            Color = ColorSequence.new({
-                ColorSequenceKeypoint.new(0, Color3.fromRGB(110, 255, 183)),
-                ColorSequenceKeypoint.new(1, Color3.fromRGB(255, 119, 253)),
-            }),
-            Rotation = 25,
-        }),
-    })
+	})
 end
 
 Button = Hooks.new(Roact)(Button)