+++ stories/hooks/useSpringEasingClick.story.lua
@@ -11,6 +11,7 @@ local GuiService = game:GetService("GuiService")
 local Roact = require(ReplicatedStorage.Packages.Roact)
 local Hooks = require(ReplicatedStorage.Packages.Hooks)
 local RoactSpring = require(ReplicatedStorage.Packages.RoactSpring)
+local CircleButton = require(script.Parent.Parent.components.CircleButton)
 
 local e = Roact.createElement
 
@@ -39,39 +40,10 @@ local function Button(props, hooks)
         end
     end)
 
-	return e("TextButton", {
-        AnchorPoint = Vector2.new(0.5, 0.5),
+	return e(CircleButton, {
         Position = styles.position,
 		Size = styles.size,
-        BackgroundColor3 = Color3.fromRGB(255, 255, 255),
-        AutoButtonColor = false,
-        Text = "",
-	}, {
-        UICorner = e("UICorner", {
-            CornerRadius = UDim.new(1, 0),
-        }),
-        UIGradient = e("UIGradient", {
-            Color = ColorSequence.new({
-                ColorSequenceKeypoint.new(0, Color3.fromRGB(110, 255, 183)),
-                ColorSequenceKeypoint.new(1, Color3.fromRGB(255, 119, 253)),
-            }),
-            Rotation = 25,
-        }),
-        UIStroke = e("UIStroke", {
-            ApplyStrokeMode = Enum.ApplyStrokeMode.Border,
-            Color = Color3.fromRGB(29, 237, 255),
-            Thickness = 5.3,
-            Transparency = 0.7,
-        }, {
-            UIGradient = e("UIGradient", {
-                Rotation = -90,
-                Transparency = NumberSequence.new({
-                    NumberSequenceKeypoint.new(0, 0),
-                    NumberSequenceKeypoint.new(1, 1),
-                }),
-            }),
-        }),
-    })
+	})
 end
 
 Button = Hooks.new(Roact)(Button)