+++ stories/others/Spring.story.lua
@@ -2,6 +2,7 @@ local ReplicatedStorage = game:GetService("ReplicatedStorage")
 
 local Roact = require(ReplicatedStorage.Packages.Roact)
 local RoactSpring = require(ReplicatedStorage.Packages.RoactSpring)
+local CircleButton = require(script.Parent.Parent.components.CircleButton)
 
 local e = Roact.createElement
 
@@ -14,21 +15,15 @@ function Example:init()
 end
 
 function Example:render()
-    return e("TextButton", {
-        AnchorPoint = Vector2.new(0.5, 0.5),
+    return e(CircleButton, {
         Position = self.styles.position,
-		Size = UDim2.fromOffset(30, 30),
-		BackgroundColor3 = Color3.fromRGB(99, 255, 130),
-        Text = "Click me",
 
         [Roact.Event.Activated] = function()
             self.api:start({
                 position = UDim2.fromScale(0.5, 0.8),
             })
         end,
-	}, {
-        UICorner = e("UICorner"),
-    })
+	})
 end
 
 return function(target)