+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Fix color3 animating with wrong values
+* Clean up all stories to use circle button component
 
 ## 0.1.1 (February 3, 2022)